kafka_consumer/requirements.txt
confluent_kafka
pymongo==4.8.0
certifi==2024.7.4
python-dotenv==0.21.1
